Delete README.md
Consumer complaint resolution is important to any business. In this particular case we have been given detailed
consumer complaints along with whether consumer disputed with the conclusion. If we are able to predict this,
consumer who is more likely to dispute a conclusion can be given more attention as to how the complaints are
handled as well as how persuasively the final conlusions are conveyed to them.
Your target here is to build prediction model for column "Consumer disputed".